Use bigint for pr_evidence_evaluation.pending_link_count
Squawk's prefer-bigint-over-int gate rejects 32-bit integer columns in
migrations; the schema file is updated to match so db diff stays clean.
Generated types are unchanged (both map to number).

+++ apps/tenant-dashboard/supabase/migrations/20260814000000_add_pr_evidence_evaluation.sql
@@ -10,7 +10,7 @@
     "pr_number" bigint not null,
     "verdict" text not null,
     "facts" jsonb not null default '[]'::jsonb,
-    "pending_link_count" integer not null default 0,
+    "pending_link_count" bigint not null default 0,
     "evaluated_at" timestamp with time zone not null default now()
       );
